requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.1.1
django-rename-app==0.1.7
django-rest-framework==0.1.0
django-sass==1.1.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djlint==1.35.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.15.1
json5==0.9.25
libsass==0.23.0
pathspec==0.12.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-magic==0.4.27
PyYAML==6.0.2
regex==2024.9.11
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
